Add support for resolving Maven artifacts
Now find_jar() function, which is used by find-jar, build-classpath,
build-jar-repository and %jpackage_script, will be able to resolve
artifacts in form:

  groupId:artifactId[:extension[:classifier]][:version]

For successfull resolution of Maven artifacts XMvn has to be
installed, but XMvn is not a hard recurement.  XMvn is not required as
long as traditional JAR names are used instead of Maven coordinates.