Do not use full paths unless needed
Traditionally some binaries are included in /bin, some in /usr/bin.
On recent versions of Fedora all binaries were moved to /usr/bin (so
called "usr move").  This is Fedora-specific.

Java Packages Tools are expected to be portable between different
systems, so we cannot assume "usr move" and we need to use binaries
from /bin or /usr/bin, as appropriate.

Even better, it's better not to use full path at all, if possible.
For example "echo" may be shell builtin, using /bin/echo would be
slower as it involves fork().