[maven.req] Handle dependecies with "UNKNOWN" version from XMvn
XMvn can generate dependency with UNKNOWN version in some cases. The package can
build but it's impossible to figure out what requires should be generated
because we don't know the origin of given dependency. It could be ordinary
package, compat package or a package with some namespace.

In that case we fail the build and enforce packager to decide if the dependency
should be removed completely or a proper BuildRequires added.