Initial implementation of %jp_binding macro
Add new experimental %jp_binding macro for generating binding
packages, which are typically used to configure Java applications such
as Maven to run with particular JVM implementation, but the mechanism
is generic and can be used for other things as well.

Unlike earlier implementations of binding packages, the new approach
does not lead to creating conflicting packages.  Instead a hierarchy
of symlinks is maintained, similarly to the alternatives system.

Currently there is no way to specify which binding will be selected in
case when there is more than one installed (there is nothing like
priorities yet), but this subject to change in the future.

The implementation of %jp_binding relies on "dynamic spec generation"
(aka "specparts") feature of RPM 4.19, but later it could be updated
to work with older RPM versions too.  More info in RPM docs:
https://rpm-software-management.github.io/rpm/manual/dynamic_specs.html